% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CF_Likelihood_pooled.R
\name{CF_Likelihood_pooled}
\alias{CF_Likelihood_pooled}
\title{Counterfactual Likelihood}
\format{
\code{\link{R6Class}} object.
}
\value{
\code{Likelihood} object
}
\description{
Represents a counterfactual likelihood where one or more likelihood factors has been replaced with an intervention as specified by \code{intervention_list}.
Inherits from \code{\link{Likelihood}}. Other factors (including their updates) are taken from an underlying \code{observed_likelihood} estimated from observed data.
}
\seealso{
Other Likelihood objects: 
\code{\link{LF_collapsed}},
\code{\link{LF_emp}},
\code{\link{LF_fit_pooled}},
\code{\link{LF_static}},
\code{\link{Targeted_Likelihood_pooled}}
}
\concept{Likelihood objects}
\keyword{data}
\section{Super classes}{
\code{\link[sl3:Lrnr_base]{sl3::Lrnr_base}} -> \code{\link[LTMLE:Likelihood_pooled]{LTMLE::Likelihood_pooled}} -> \code{CF_Likelihood_pooled}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{CF_Likelihood_pooled$new()}}
\item \href{#method-enumerate_cf_tasks}{\code{CF_Likelihood_pooled$enumerate_cf_tasks()}}
\item \href{#method-get_likelihood}{\code{CF_Likelihood_pooled$get_likelihood()}}
\item \href{#method-clone}{\code{CF_Likelihood_pooled$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="assert_trained">}\href{../../sl3/html/Lrnr_base.html#method-assert_trained}{\code{sl3::Lrnr_base$assert_trained()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="base_chain">}\href{../../sl3/html/Lrnr_base.html#method-base_chain}{\code{sl3::Lrnr_base$base_chain()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="base_predict">}\href{../../sl3/html/Lrnr_base.html#method-base_predict}{\code{sl3::Lrnr_base$base_predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="chain">}\href{../../sl3/html/Lrnr_base.html#method-chain}{\code{sl3::Lrnr_base$chain()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="custom_chain">}\href{../../sl3/html/Lrnr_base.html#method-custom_chain}{\code{sl3::Lrnr_base$custom_chain()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="get_outcome_range">}\href{../../sl3/html/Lrnr_base.html#method-get_outcome_range}{\code{sl3::Lrnr_base$get_outcome_range()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="get_outcome_type">}\href{../../sl3/html/Lrnr_base.html#method-get_outcome_type}{\code{sl3::Lrnr_base$get_outcome_type()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="predict">}\href{../../sl3/html/Lrnr_base.html#method-predict}{\code{sl3::Lrnr_base$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="predict_fold">}\href{../../sl3/html/Lrnr_base.html#method-predict_fold}{\code{sl3::Lrnr_base$predict_fold()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="set_train">}\href{../../sl3/html/Lrnr_base.html#method-set_train}{\code{sl3::Lrnr_base$set_train()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="subset_covariates">}\href{../../sl3/html/Lrnr_base.html#method-subset_covariates}{\code{sl3::Lrnr_base$subset_covariates()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="train">}\href{../../sl3/html/Lrnr_base.html#method-train}{\code{sl3::Lrnr_base$train()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="train_sublearners">}\href{../../sl3/html/Lrnr_base.html#method-train_sublearners}{\code{sl3::Lrnr_base$train_sublearners()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="LTMLE" data-topic="Likelihood_pooled" data-id="add_factors">}\href{../../LTMLE/html/Likelihood_pooled.html#method-add_factors}{\code{LTMLE::Likelihood_pooled$add_factors()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="LTMLE" data-topic="Likelihood_pooled" data-id="base_train">}\href{../../LTMLE/html/Likelihood_pooled.html#method-base_train}{\code{LTMLE::Likelihood_pooled$base_train()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="LTMLE" data-topic="Likelihood_pooled" data-id="get_likelihoods">}\href{../../LTMLE/html/Likelihood_pooled.html#method-get_likelihoods}{\code{LTMLE::Likelihood_pooled$get_likelihoods()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="LTMLE" data-topic="Likelihood_pooled" data-id="get_possible_counterfactuals">}\href{../../LTMLE/html/Likelihood_pooled.html#method-get_possible_counterfactuals}{\code{LTMLE::Likelihood_pooled$get_possible_counterfactuals()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="LTMLE" data-topic="Likelihood_pooled" data-id="print">}\href{../../LTMLE/html/Likelihood_pooled.html#method-print}{\code{LTMLE::Likelihood_pooled$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="LTMLE" data-topic="Likelihood_pooled" data-id="sample">}\href{../../LTMLE/html/Likelihood_pooled.html#method-sample}{\code{LTMLE::Likelihood_pooled$sample()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="LTMLE" data-topic="Likelihood_pooled" data-id="validate_task">}\href{../../LTMLE/html/Likelihood_pooled.html#method-validate_task}{\code{LTMLE::Likelihood_pooled$validate_task()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CF_Likelihood_pooled$new(observed_likelihood, intervention_list, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-enumerate_cf_tasks"></a>}}
\if{latex}{\out{\hypertarget{method-enumerate_cf_tasks}{}}}
\subsection{Method \code{enumerate_cf_tasks()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CF_Likelihood_pooled$enumerate_cf_tasks(tmle_task)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_likelihood"></a>}}
\if{latex}{\out{\hypertarget{method-get_likelihood}{}}}
\subsection{Method \code{get_likelihood()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CF_Likelihood_pooled$get_likelihood(
  tmle_task,
  node,
  fold_number = "full",
  drop_id = F,
  to_wide = F
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{CF_Likelihood_pooled$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
