% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LF_fit_pooled.R
\name{LF_fit_pooled}
\alias{LF_fit_pooled}
\title{Likelihood Factor Estimated from Data using sl3.}
\format{
\code{\link{R6Class}} object.
}
\value{
\code{LF_base} object
}
\description{
Uses an \code{sl3} learner to estimate a likelihood factor from data.
Similar to LF_fit object but allows vector-valued names (corresponding to multiple nodes sharing the same regression_task variables but at different times)
This learner will pool the regression tasks across time. This depends on ltmle3_Task which has the pooled task functionality.
Inherits from \code{\link{LF_base}}; see that page for documentation on likelihood factors in general.
}
\section{Constructor}{

  \code{define_lf(LF_fit, name, learner, ..., type = "density")}

  \describe{
    \item{\code{name}}{character, the name of the factor. Should match a node name in the nodes specified by \code{\link{tmle3_Task}$npsem}
    }
    \item{\code{learner}}{An sl3 learner to be used to estimate the factor
    }
    \item{\code{...}}{Not currently used.
    }
    \item{\code{type}}{character, either "density", for conditional density or, "mean" for conditional mean
    }
    }
}

\section{Fields}{

\describe{
    \item{\code{learner}}{The learner or learner fit object}
    }
TODO If outcome is multivariate then this assumes learner will handle the multi outcome and return predictions as a matrix
}

\seealso{
Other Likelihood objects: 
\code{\link{CF_Likelihood_pooled}},
\code{\link{LF_collapsed}},
\code{\link{LF_emp}},
\code{\link{LF_static}},
\code{\link{Targeted_Likelihood_pooled}}
}
\concept{Likelihood objects}
\keyword{data}
\section{Super class}{
\code{\link[tmle3:LF_base]{tmle3::LF_base}} -> \code{LF_fit_pooled}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{LF_fit_pooled$new()}}
\item \href{#method-delayed_train}{\code{LF_fit_pooled$delayed_train()}}
\item \href{#method-train}{\code{LF_fit_pooled$train()}}
\item \href{#method-shape_predictions}{\code{LF_fit_pooled$shape_predictions()}}
\item \href{#method-get_mean}{\code{LF_fit_pooled$get_mean()}}
\item \href{#method-get_density}{\code{LF_fit_pooled$get_density()}}
\item \href{#method-get_likelihood}{\code{LF_fit_pooled$get_likelihood()}}
\item \href{#method-sample}{\code{LF_fit_pooled$sample()}}
\item \href{#method-clone}{\code{LF_fit_pooled$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="LF_base" data-id="cf_values">}\href{../../tmle3/html/LF_base.html#method-cf_values}{\code{tmle3::LF_base$cf_values()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="LF_base" data-id="print">}\href{../../tmle3/html/LF_base.html#method-print}{\code{tmle3::LF_base$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_fit_pooled$new(
  name,
  learner,
  is_time_variant = FALSE,
  ...,
  type = "density"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delayed_train"></a>}}
\if{latex}{\out{\hypertarget{method-delayed_train}{}}}
\subsection{Method \code{delayed_train()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_fit_pooled$delayed_train(tmle_task)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-train"></a>}}
\if{latex}{\out{\hypertarget{method-train}{}}}
\subsection{Method \code{train()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_fit_pooled$train(tmle_task, learner_fit)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-shape_predictions"></a>}}
\if{latex}{\out{\hypertarget{method-shape_predictions}{}}}
\subsection{Method \code{shape_predictions()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_fit_pooled$shape_predictions(learner_task, preds)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_mean"></a>}}
\if{latex}{\out{\hypertarget{method-get_mean}{}}}
\subsection{Method \code{get_mean()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_fit_pooled$get_mean(tmle_task, fold_number)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_density"></a>}}
\if{latex}{\out{\hypertarget{method-get_density}{}}}
\subsection{Method \code{get_density()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_fit_pooled$get_density(
  tmle_task,
  fold_number,
  check_at_risk = T,
  to_mat = T
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_likelihood"></a>}}
\if{latex}{\out{\hypertarget{method-get_likelihood}{}}}
\subsection{Method \code{get_likelihood()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_fit_pooled$get_likelihood(tmle_task, fold_number = "full")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample"></a>}}
\if{latex}{\out{\hypertarget{method-sample}{}}}
\subsection{Method \code{sample()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_fit_pooled$sample(tmle_task, n_samples = NULL, fold_number = "full")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LF_fit_pooled$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
