% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihood_pooled.R
\name{Likelihood_pooled}
\alias{Likelihood_pooled}
\title{Likelihood object that supports a list of pooled LF objects (e.g. LF_fit_pooled)
All output behavior is identical to the standard Likelihood object.}
\description{
Likelihood object that supports a list of pooled LF objects (e.g. LF_fit_pooled)
All output behavior is identical to the standard Likelihood object.

Likelihood object that supports a list of pooled LF objects (e.g. LF_fit_pooled)
All output behavior is identical to the standard Likelihood object.
}
\section{Super classes}{
\code{\link[sl3:Lrnr_base]{sl3::Lrnr_base}} -> \code{\link[tmle3:Likelihood]{tmle3::Likelihood}} -> \code{Likelihood_pooled}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Likelihood_pooled$new()}}
\item \href{#method-validate_task}{\code{Likelihood_pooled$validate_task()}}
\item \href{#method-get_likelihood}{\code{Likelihood_pooled$get_likelihood()}}
\item \href{#method-get_likelihoods}{\code{Likelihood_pooled$get_likelihoods()}}
\item \href{#method-clone}{\code{Likelihood_pooled$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="assert_trained">}\href{../../sl3/html/Lrnr_base.html#method-assert_trained}{\code{sl3::Lrnr_base$assert_trained()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="base_chain">}\href{../../sl3/html/Lrnr_base.html#method-base_chain}{\code{sl3::Lrnr_base$base_chain()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="base_predict">}\href{../../sl3/html/Lrnr_base.html#method-base_predict}{\code{sl3::Lrnr_base$base_predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="chain">}\href{../../sl3/html/Lrnr_base.html#method-chain}{\code{sl3::Lrnr_base$chain()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="custom_chain">}\href{../../sl3/html/Lrnr_base.html#method-custom_chain}{\code{sl3::Lrnr_base$custom_chain()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="get_outcome_range">}\href{../../sl3/html/Lrnr_base.html#method-get_outcome_range}{\code{sl3::Lrnr_base$get_outcome_range()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="get_outcome_type">}\href{../../sl3/html/Lrnr_base.html#method-get_outcome_type}{\code{sl3::Lrnr_base$get_outcome_type()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="predict">}\href{../../sl3/html/Lrnr_base.html#method-predict}{\code{sl3::Lrnr_base$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="predict_fold">}\href{../../sl3/html/Lrnr_base.html#method-predict_fold}{\code{sl3::Lrnr_base$predict_fold()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="set_train">}\href{../../sl3/html/Lrnr_base.html#method-set_train}{\code{sl3::Lrnr_base$set_train()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="subset_covariates">}\href{../../sl3/html/Lrnr_base.html#method-subset_covariates}{\code{sl3::Lrnr_base$subset_covariates()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="train">}\href{../../sl3/html/Lrnr_base.html#method-train}{\code{sl3::Lrnr_base$train()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="train_sublearners">}\href{../../sl3/html/Lrnr_base.html#method-train_sublearners}{\code{sl3::Lrnr_base$train_sublearners()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="Likelihood" data-id="add_factors">}\href{../../tmle3/html/Likelihood.html#method-add_factors}{\code{tmle3::Likelihood$add_factors()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="Likelihood" data-id="base_train">}\href{../../tmle3/html/Likelihood.html#method-base_train}{\code{tmle3::Likelihood$base_train()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="Likelihood" data-id="get_possible_counterfactuals">}\href{../../tmle3/html/Likelihood.html#method-get_possible_counterfactuals}{\code{tmle3::Likelihood$get_possible_counterfactuals()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="Likelihood" data-id="print">}\href{../../tmle3/html/Likelihood.html#method-print}{\code{tmle3::Likelihood$print()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="tmle3" data-topic="Likelihood" data-id="sample">}\href{../../tmle3/html/Likelihood.html#method-sample}{\code{tmle3::Likelihood$sample()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Likelihood_pooled$new(factor_list, cache = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate_task"></a>}}
\if{latex}{\out{\hypertarget{method-validate_task}{}}}
\subsection{Method \code{validate_task()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Likelihood_pooled$validate_task(tmle_task)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_likelihood"></a>}}
\if{latex}{\out{\hypertarget{method-get_likelihood}{}}}
\subsection{Method \code{get_likelihood()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Likelihood_pooled$get_likelihood(tmle_task, node, fold_number = "full")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_likelihoods"></a>}}
\if{latex}{\out{\hypertarget{method-get_likelihoods}{}}}
\subsection{Method \code{get_likelihoods()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Likelihood_pooled$get_likelihoods(
  tmle_task,
  nodes = NULL,
  fold_number = "full"
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Likelihood_pooled$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
