% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Params_skeleton.R
\name{Param_base}
\alias{Param_base}
\title{Base Class for Defining Parameters}
\format{
\code{\link{R6Class}} object.
}
\value{
\code{Param_base} object
}
\description{
A parameter is a function of the likelihood. Once given a \code{\link{Likelihood}} object, a parameter will a value.
These objects also contain information about the efficient influence function (EIF) of a parameter, as well as its clever covariate(s).
}
\seealso{
Other Parameters: 
\code{\link{Param_survival}}
}
\concept{Parameters}
\keyword{data}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Param_base$new()}}
\item \href{#method-clever_covariates}{\code{Param_base$clever_covariates()}}
\item \href{#method-submodel_info}{\code{Param_base$submodel_info()}}
\item \href{#method-loss_function}{\code{Param_base$loss_function()}}
\item \href{#method-submodel}{\code{Param_base$submodel()}}
\item \href{#method-estimates}{\code{Param_base$estimates()}}
\item \href{#method-print}{\code{Param_base$print()}}
\item \href{#method-clone}{\code{Param_base$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_base$new(observed_likelihood, ..., nodes_to_target = "Y")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clever_covariates"></a>}}
\if{latex}{\out{\hypertarget{method-clever_covariates}{}}}
\subsection{Method \code{clever_covariates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_base$clever_covariates(tmle_task, fold_number = "full")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-submodel_info"></a>}}
\if{latex}{\out{\hypertarget{method-submodel_info}{}}}
\subsection{Method \code{submodel_info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_base$submodel_info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loss_function"></a>}}
\if{latex}{\out{\hypertarget{method-loss_function}{}}}
\subsection{Method \code{loss_function()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_base$loss_function(estimate_list, observed_list)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-submodel"></a>}}
\if{latex}{\out{\hypertarget{method-submodel}{}}}
\subsection{Method \code{submodel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_base$submodel(epsilon, initial_list, H_list)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimates"></a>}}
\if{latex}{\out{\hypertarget{method-estimates}{}}}
\subsection{Method \code{estimates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_base$estimates(tmle_task = NULL, fold_number = "full")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_base$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_base$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
