% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Param_survival.R
\name{Param_survival}
\alias{Param_survival}
\title{Survival Curve}
\format{
\code{\link{R6Class}} object.
}
\value{
\code{Param_base} object
}
\description{
Survival Curve

Survival Curve
}
\section{Constructor}{

  \code{define_param(Param_survival, observed_likelihood, intervention_list, ..., outcome_node)}

  \describe{
    \item{\code{observed_likelihood}}{A \code{\link{Likelihood}} corresponding to the observed likelihood
    }
    \item{\code{intervention_list}}{A list of objects inheriting from \code{\link{LF_base}}, representing the intervention.
    }
    \item{\code{...}}{Not currently used.
    }
    \item{\code{outcome_node}}{character, the name of the node that should be treated as the outcome
    }
    }
}

\section{Fields}{

\describe{
    \item{\code{cf_likelihood}}{the counterfactual likelihood for this treatment
    }
    \item{\code{intervention_list}}{A list of objects inheriting from \code{\link{LF_base}}, representing the intervention
    }
}
}

\seealso{
Other Parameters: 
\code{\link{Param_base}},
\code{\link{define_param}()}
}
\concept{Parameters}
\keyword{data}
\section{Super class}{
\code{\link[LTMLE:Param_base]{LTMLE::Param_base}} -> \code{Param_survival}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Param_survival$new()}}
\item \href{#method-long_to_mat}{\code{Param_survival$long_to_mat()}}
\item \href{#method-hm_to_sm}{\code{Param_survival$hm_to_sm()}}
\item \href{#method-clever_covariates_internal}{\code{Param_survival$clever_covariates_internal()}}
\item \href{#method-submodel_info}{\code{Param_survival$submodel_info()}}
\item \href{#method-offset_transform}{\code{Param_survival$offset_transform()}}
\item \href{#method-submodel_family}{\code{Param_survival$submodel_family()}}
\item \href{#method-submodel}{\code{Param_survival$submodel()}}
\item \href{#method-loss_function}{\code{Param_survival$loss_function()}}
\item \href{#method-clever_covariates}{\code{Param_survival$clever_covariates()}}
\item \href{#method-get_EIC_component}{\code{Param_survival$get_EIC_component()}}
\item \href{#method-estimates}{\code{Param_survival$estimates()}}
\item \href{#method-clone}{\code{Param_survival$clone()}}
}
}
\if{html}{
\out{<details open ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="LTMLE" data-topic="Param_base" data-id="print">}\href{../../LTMLE/html/Param_base.html#method-print}{\code{LTMLE::Param_base$print()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_survival$new(
  observed_likelihood,
  intervention_list,
  ...,
  outcome_node,
  target_times = NULL
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-long_to_mat"></a>}}
\if{latex}{\out{\hypertarget{method-long_to_mat}{}}}
\subsection{Method \code{long_to_mat()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_survival$long_to_mat(x, id, time)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-hm_to_sm"></a>}}
\if{latex}{\out{\hypertarget{method-hm_to_sm}{}}}
\subsection{Method \code{hm_to_sm()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_survival$hm_to_sm(hm)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clever_covariates_internal"></a>}}
\if{latex}{\out{\hypertarget{method-clever_covariates_internal}{}}}
\subsection{Method \code{clever_covariates_internal()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_survival$clever_covariates_internal(
  tmle_task = NULL,
  fold_number = "full",
  subset_times = FALSE,
  for_fitting = T
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-submodel_info"></a>}}
\if{latex}{\out{\hypertarget{method-submodel_info}{}}}
\subsection{Method \code{submodel_info()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_survival$submodel_info()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-offset_transform"></a>}}
\if{latex}{\out{\hypertarget{method-offset_transform}{}}}
\subsection{Method \code{offset_transform()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_survival$offset_transform(initial, observed)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-submodel_family"></a>}}
\if{latex}{\out{\hypertarget{method-submodel_family}{}}}
\subsection{Method \code{submodel_family()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_survival$submodel_family()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-submodel"></a>}}
\if{latex}{\out{\hypertarget{method-submodel}{}}}
\subsection{Method \code{submodel()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_survival$submodel(epsilon, initial, H, observed)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-loss_function"></a>}}
\if{latex}{\out{\hypertarget{method-loss_function}{}}}
\subsection{Method \code{loss_function()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_survival$loss_function(estimate, observed)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clever_covariates"></a>}}
\if{latex}{\out{\hypertarget{method-clever_covariates}{}}}
\subsection{Method \code{clever_covariates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_survival$clever_covariates(
  tmle_task,
  fold_number = "full",
  for_fitting = T
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_EIC_component"></a>}}
\if{latex}{\out{\hypertarget{method-get_EIC_component}{}}}
\subsection{Method \code{get_EIC_component()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_survival$get_EIC_component(task, node)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-estimates"></a>}}
\if{latex}{\out{\hypertarget{method-estimates}{}}}
\subsection{Method \code{estimates()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_survival$estimates(tmle_task = NULL, fold_number = "full")}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Param_survival$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
