\name{PreviousReleaseTests}
\alias{btests}
\docType{data}
\title{
Sample data for testing
}
\description{
Sample data and results generated by a previous version of ltmle. The data is only
for testing purposes. It is not intended for most users.
}
\usage{data(PreviousReleaseTests)}
\format{
  list of lists with components: fun, args, compareFunction, info, result
}
\source{
simulated data
}

\examples{
data(PreviousReleaseTests)
}
\keyword{datasets}
