% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Likelihood_pooled.R
\name{Likelihood_pooled}
\alias{Likelihood_pooled}
\title{Likelihood object that supports a list of pooled LF objects (e.g. LF_fit_pooled)
All output behavior is identical to the standard Likelihood object.}
\description{
Likelihood object that supports a list of pooled LF objects (e.g. LF_fit_pooled)
All output behavior is identical to the standard Likelihood object.

Likelihood object that supports a list of pooled LF objects (e.g. LF_fit_pooled)
All output behavior is identical to the standard Likelihood object.
}
\section{Super class}{
\code{\link[sl3:Lrnr_base]{sl3::Lrnr_base}} -> \code{Likelihood_pooled}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Likelihood_pooled$new()}}
\item \href{#method-print}{\code{Likelihood_pooled$print()}}
\item \href{#method-validate_task}{\code{Likelihood_pooled$validate_task()}}
\item \href{#method-get_likelihood}{\code{Likelihood_pooled$get_likelihood()}}
\item \href{#method-get_likelihoods}{\code{Likelihood_pooled$get_likelihoods()}}
\item \href{#method-get_possible_counterfactuals}{\code{Likelihood_pooled$get_possible_counterfactuals()}}
\item \href{#method-base_train}{\code{Likelihood_pooled$base_train()}}
\item \href{#method-add_factors}{\code{Likelihood_pooled$add_factors()}}
\item \href{#method-sample}{\code{Likelihood_pooled$sample()}}
\item \href{#method-clone}{\code{Likelihood_pooled$clone()}}
}
}
\if{html}{
\out{<details ><summary>Inherited methods</summary>}
\itemize{
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="assert_trained">}\href{../../sl3/html/Lrnr_base.html#method-assert_trained}{\code{sl3::Lrnr_base$assert_trained()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="base_chain">}\href{../../sl3/html/Lrnr_base.html#method-base_chain}{\code{sl3::Lrnr_base$base_chain()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="base_predict">}\href{../../sl3/html/Lrnr_base.html#method-base_predict}{\code{sl3::Lrnr_base$base_predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="chain">}\href{../../sl3/html/Lrnr_base.html#method-chain}{\code{sl3::Lrnr_base$chain()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="custom_chain">}\href{../../sl3/html/Lrnr_base.html#method-custom_chain}{\code{sl3::Lrnr_base$custom_chain()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="get_outcome_range">}\href{../../sl3/html/Lrnr_base.html#method-get_outcome_range}{\code{sl3::Lrnr_base$get_outcome_range()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="get_outcome_type">}\href{../../sl3/html/Lrnr_base.html#method-get_outcome_type}{\code{sl3::Lrnr_base$get_outcome_type()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="predict">}\href{../../sl3/html/Lrnr_base.html#method-predict}{\code{sl3::Lrnr_base$predict()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="predict_fold">}\href{../../sl3/html/Lrnr_base.html#method-predict_fold}{\code{sl3::Lrnr_base$predict_fold()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="set_train">}\href{../../sl3/html/Lrnr_base.html#method-set_train}{\code{sl3::Lrnr_base$set_train()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="subset_covariates">}\href{../../sl3/html/Lrnr_base.html#method-subset_covariates}{\code{sl3::Lrnr_base$subset_covariates()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="train">}\href{../../sl3/html/Lrnr_base.html#method-train}{\code{sl3::Lrnr_base$train()}}\out{</span>}
\item \out{<span class="pkg-link" data-pkg="sl3" data-topic="Lrnr_base" data-id="train_sublearners">}\href{../../sl3/html/Lrnr_base.html#method-train_sublearners}{\code{sl3::Lrnr_base$train_sublearners()}}\out{</span>}
}
\out{</details>}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Likelihood_pooled$new(factor_list, cache = NULL, ...)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Likelihood_pooled$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-validate_task"></a>}}
\if{latex}{\out{\hypertarget{method-validate_task}{}}}
\subsection{Method \code{validate_task()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Likelihood_pooled$validate_task(tmle_task)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_likelihood"></a>}}
\if{latex}{\out{\hypertarget{method-get_likelihood}{}}}
\subsection{Method \code{get_likelihood()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Likelihood_pooled$get_likelihood(
  tmle_task,
  node,
  fold_number = "full",
  drop_id = F,
  to_wide = F
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_likelihoods"></a>}}
\if{latex}{\out{\hypertarget{method-get_likelihoods}{}}}
\subsection{Method \code{get_likelihoods()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Likelihood_pooled$get_likelihoods(
  tmle_task,
  nodes = NULL,
  fold_number = "full",
  drop_id = F,
  to_wide = F
)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_possible_counterfactuals"></a>}}
\if{latex}{\out{\hypertarget{method-get_possible_counterfactuals}{}}}
\subsection{Method \code{get_possible_counterfactuals()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Likelihood_pooled$get_possible_counterfactuals(nodes = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-base_train"></a>}}
\if{latex}{\out{\hypertarget{method-base_train}{}}}
\subsection{Method \code{base_train()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Likelihood_pooled$base_train(task, pretrain)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_factors"></a>}}
\if{latex}{\out{\hypertarget{method-add_factors}{}}}
\subsection{Method \code{add_factors()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Likelihood_pooled$add_factors(factor_list)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-sample"></a>}}
\if{latex}{\out{\hypertarget{method-sample}{}}}
\subsection{Method \code{sample()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Likelihood_pooled$sample(tmle_task = NULL, sample_lib = NULL)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\if{latex}{\out{\hypertarget{method-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Likelihood_pooled$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
