% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/ltmle-package.R
\docType{data}
\name{PreviousReleaseTests}
\alias{PreviousReleaseTests}
\alias{btests}
\title{Sample data for testing}
\format{list of lists with components: fun, args, compareFunction, info,
result}
\source{
simulated data
}
\description{
Sample data and results generated by a previous version of ltmle. The data
is only for testing purposes. It is not intended for most users.
}
\examples{
data(PreviousReleaseTests)
}
\keyword{datasets}

